requirements.txt
alembic==1.2.1
apturl==0.5.2
asgiref==3.2.3
asn1crypto==0.24.0
astroid==2.3.2
autopep8==1.4.4
beautifulsoup4==4.8.2
blinker==1.4
Brlapi==0.6.6
certifi==2018.1.18
chardet==3.0.4
Click==7.0
colorama==0.3.7
command-not-found==0.3
confusable-homoglyphs==3.2.0
cryptography==2.1.4
cupshelpers==1.0
defer==1.0.6
distro-info===0.18ubuntu0.18.04.1
dj-database-url==0.5.0
Django==3.0.3
django-admin-colors==0.2.0
django-authtools==1.7.0
django-bootstrap-form==3.4
django-bootstrap3==11.1.0
django-bootstrap4==1.1.1
django-crispy-forms==1.8.1
django-filter==2.2.0
django-heroku==0.3.1
django-online-users==0.3
django-registration==3.1
django-summernote==0.8.11.6
django-tinymce==2.8.0
gunicorn==20.0.4
httplib2==0.9.2
idna==2.6
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.10.3
kazam==1.4.5
keyring==10.6.0
keyrings.alt==3.0
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
lazy-object-proxy==1.4.3
louis==3.5.0
macaroonbakery==1.1.3
Mako==1.1.0
MarkupSafe==1.1.1
mccabe==0.6.1
netifaces==0.10.4
oauth==1.0.1
olefile==0.45.1
pep8==1.7.1
pexpect==4.2.1
Pillow==5.1.0
protobuf==3.0.0
psutil==5.4.2
psycopg2==2.8.4
pycairo==1.16.2
PyChromecast==0.8.1
pycodestyle==2.5.0
pycrypto==2.6.1
pycups==1.9.73
pyfiglet==0.7.4
pygobject==3.26.1
pyinotify==0.9.6
pylint==2.4.3
pymacaroons==0.13.0
PyNaCl==1.1.2
pyOpenSSL==17.5.0
pyperclip==1.7.0
pyRFC3339==1.0
python-apt==1.6.5+ubuntu0.2
python-dateutil==2.8.0
python-debian==0.1.32
python-decouple==3.3
python-editor==1.0.4
pytz==2019.3
pyxattr==0.6.0
pyxdg==0.25
PyYAML==3.12
reportlab==3.4.0
requests==2.18.4
requests-unixsocket==0.1.5
SecretStorage==2.3.1
simplejson==3.13.2
six==1.12.0
soupsieve==1.9.5
SQLAlchemy==1.3.10
sqlparse==0.3.0
system-service==0.3
systemd-python==234
typed-ast==1.4.0
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.22
usb-creator==0.3.3
virtualenv==15.1.0
wadllib==1.3.2
Werkzeug==0.16.0
whitenoise==5.0.1
wrapt==1.11.2
xkit==0.0.0
youtube-dl==2018.3.14
zeroconf==0.19.1
zope.interface==4.3.2
